Refactored synthesis, but there is a problem with theBeat. About to move theBeat to its own module
